requirements.txt
Flask==2.2.5
gunicorn==20.1.0
python-dotenv
Flask-RESTX==1.1.0
Werkzeug==2.2.2
PyJWT==2.8.0
marshmallow==3.20.1
email-validator==2.0.0
